# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Yume Ramen Noodles API",
    "version": "1.0.0",
    "description": "API documentation for Yume Ramen Noodles"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
  "paths": {
...
    "/dishes": {
      "get": {
        "summary": "Returns the list of all dishes",
        "tags": [
          "Dishes"
        ],
        "responses": {
          "200": {
            "description": "The list of the dishes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Dish"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Create a new dish",
        "tags": [
          "Dishes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Requires Admin privileges",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "Name",
                  "Price",
                  "Ingredients"
                ],
                "properties": {
                  "Name": {
                    "type": "string"
                  },
                  "Price": {
                    "type": "number",
                    "format": "float"
                  },
                  "Ingredients": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The dish was successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dish"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/dishes/{id}": {
      "get": {
        "summary": "Get the dish by id",
        "tags": [
          "Dishes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "The dish id"
          }
        ],
        "responses": {
          "200": {
            "description": "The dish description by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dish"
                }
              }
            }
          },
          "404": {
            "description": "The dish was not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "summary": "Update the dish by the id",
        "tags": [
          "Dishes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Requires Admin privileges",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "The dish id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Name": {
                    "type": "string"
                  },
                  "Price": {
                    "type": "number",
                    "format": "float"
                  },
                  "Ingredients": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The dish was updated"
          },
          "404": {
            "description": "The dish was not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "Remove the dish by id",
        "tags": [
          "Dishes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Requires Admin privileges",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "The dish id"
          }
        ],
        "responses": {
          "200": {
            "description": "The dish was deleted"
          },
          "404": {
            "description": "The dish was not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/orders": {
      "get": {
        "summary": "Returns the list of all orders",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The list of the orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Create a new order",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "items"
                ],
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "dishID"
                      ],
                      "properties": {
                        "dishID": {
                          "type": "integer"
                        },
                        "quantity": {
                          "type": "integer",
                          "default": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The order was successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "orderId": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "summary": "Get the order by id",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "The order id"
          }
        ],
        "responses": {
          "200": {
            "description": "The order description by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "The order was not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "summary": "Update the order by the id",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "The order id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Status": {
                    "type": "string",
                    "enum": [
                      "ordered",
                      "processing",
                      "delivering",
                      "completed"
                    ]
                  },
                  "Paid": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The order was updated"
          },
          "404": {
            "description": "The order was not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "Remove the order by id",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "The order id"
          }
        ],
        "responses": {
          "200": {
            "description": "The order was deleted"
          },
          "404": {
            "description": "The order was not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Dish": {
        "type": "object",
        "required": [
          "Name",
          "Price",
          "Ingredients"
        ],
        "properties": {
          "DishID": {
            "type": "integer",
            "description": "The auto-generated id of the dish"
          },
          "Name": {
            "type": "string",
            "description": "The name of the dish"
          },
          "Price": {
            "type": "number",
            "format": "float",
            "description": "The price of the dish"
          },
          "Ingredients": {
            "type": "string",
            "description": "The ingredients of the dish"
          },
          "Image": {
            "type": "string",
            "description": "The URL path to the dish image"
          }
        },
        "example": {
          "DishID": 1,
          "Name": "Tonkotsu Ramen",
          "Price": 12.5,
          "Ingredients": "Pork, Noodles, Egg, Green Onion",
          "Image": "uploads/image-123456789.jpg"
        }
      },
      "Order": {
        "type": "object",
        "required": [
          "UserID"
        ],
        "properties": {
          "OrderID": {
            "type": "integer",
            "description": "The auto-generated id of the order"
          },
          "UserID": {
            "type": "integer",
            "description": "The id of the user who placed the order"
          },
          "Ordered_at": {
            "type": "string",
            "format": "date-time",
            "description": "The date the order was placed"
          },
          "processing_at": {
            "type": "string",
            "format": "date-time"
          },
          "Delivering_at": {
            "type": "string",
            "format": "date-time"
          },
          "Completed_at": {
            "type": "string",
            "format": "date-time"
          },
          "Paid": {
            "type": "boolean",
            "description": "Whether the order has been paid"
          },
          "Status": {
            "type": "string",
            "enum": [
              "ordered",
              "processing",
              "delivering",
              "completed"
            ],
            "description": "The status of the order"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "dishID": {
                  "type": "integer"
                },
                "quantity": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                },
                "price": {
                  "type": "number"
                }
              }
            }
          }
        },
        "example": {
          "OrderID": 1,
          "UserID": 1,
          "Ordered_at": "2023-10-27T10:00:00Z",
          "Paid": false,
          "Status": "ordered",
          "items": [
            {
              "dishID": 2,
              "quantity": 1,
              "name": "Miso Ramen",
              "price": 13.5
            }
          ]
        }
      },
...
    }
  },
...
}